Add a Dockerfile for Laravel
Based on the `php:<version>-fpm` image, this brings over the application
code and any additional required dependencies.

As noted on the upstream README, we take the production `php.ini` as
this is intended for hosting. We install the current LTS of Node.js, and
bring `composer` over from it's most recent Docker image.

Finally, we run `php artisan optimize` to cache configuration, events,
routes and views.

It can be built and run with:

    docker build . -t nickcharlton/php-k8s-laravel
    docker run -it --rm  -p 9000:9000 nickcharlton/php-k8s-laravel

And tested with:

    SCRIPT_FILENAME=/var/www/html/public/index.php \
    REQUEST_METHOD=GET \
    REQUEST_URI=/up \
    cgi-fcgi -bind -connect localhost:9000

https://hub.docker.com/_/php/
https://laravel.com/docs/12.x/deployment